Move over enum.
